feat: Move WebGLProgramParameters to WebGLPrograms
